Entity IDs changed to Long values. Main menu redesigned. New entities: channel, group etc.

+++ VdrChannelManager/src/main/java/di/vdrchman/controller/TranspondersBacking.java
@@ -143,7 +143,7 @@ public class TranspondersBacking {
 
 	public void onSourceMenuSelection(ValueChangeEvent event) {
 		Transponder lastPageTopTransponder;
-		int filteredSourceId;
+		long filteredSourceId;
 		Transponder copiedTransponder;
 
 		lastPageTopTransponder = null;
@@ -152,7 +152,7 @@ public class TranspondersBacking {
 					(transpondersManager.getScrollerPage() - 1)
 							* transpondersManager.getRowsPerPage());
 		}
-		filteredSourceId = (Integer) event.getNewValue();
+		filteredSourceId = (Long) event.getNewValue();
 		if (filteredSourceId >= 0) {
 			copiedTransponder = transpondersManager.getCopiedTransponder();
 			if (copiedTransponder != null) {
